requirements.txt
configobj>=5.0.6
configparser>=3.5.0
ephem>=3.7.6.0
future>=0.15.2
matplotlib>=1.5.3
numpy>=1.6.0,!=1.15.0
pandas>=0.24.0
pillow>=4.0.0
requests>=2.11.1
scandir>=1.4

setuptools~=57.0.0
laspy~=2.5.0
